refactor: isolate dashboard automation runs
