+++ src/dashboard/mod.rs
@@ -26,6 +26,7 @@ pub(crate) mod assets;
 mod automation_config_api;
 mod automation_fact_proposals_api;
 mod automation_run_api;
+mod automation_run_service;
 mod automation_scheduler_api;
 mod automation_skills_api;
 mod curate_preview_store;
